+++ pom.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
 
@@ -12,10 +13,10 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <kotlin.version>1.1.1</kotlin.version>
+        <kotlin.version>1.3.11</kotlin.version>
         <junit.version>4.12</junit.version>
-        <vertx.version>3.4.1</vertx.version>
-        <kotlinx.coroutines.version>0.12</kotlinx.coroutines.version>
+        <vertx.version>3.6.2</vertx.version>
+        <kotlinx.coroutines.version>1.1.0</kotlinx.coroutines.version>
     </properties>
 
     <repositories>
