+++ src/main/kotlin/cn/codetector/vertx/jdbc/coroutine/JDBCClientExt.kt
@@ -2,10 +2,12 @@ package cn.codetector.vertx.jdbc.coroutine
 
 import io.vertx.ext.jdbc.JDBCClient
 import io.vertx.ext.sql.SQLConnection
-import kotlin.coroutines.experimental.suspendCoroutine
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
+import kotlin.coroutines.suspendCoroutine
 
 suspend fun JDBCClient.getConnection(): SQLConnection = suspendCoroutine { cont ->
-    this.getConnection{ conn ->
+    this.getConnection { conn ->
         if (conn.succeeded()) {
             cont.resume(conn.result())
         } else {
